Adding types.ts file
