+++ app/routes/index.tsx
@@ -5,62 +5,7 @@ import {
   getTrendingProducts,
 } from "~/models/collection.client";
 import { useState, useEffect } from "react";
-
-export interface CollectionObj {
-  key: string;
-  name: string;
-  imgSrc: string;
-  altTxt: string | undefined;
-}
-
-export interface Collection {
-  node: {
-    title: string;
-    products: {
-      edges: [
-        {
-          node: {
-            id: string;
-            images: {
-              edges: [
-                {
-                  node: {
-                    altText: string | undefined;
-                    src: string;
-                  };
-                }
-              ];
-            };
-          };
-        }
-      ];
-    };
-  };
-}
-
-export interface ProductObj {
-  key: string;
-  name: string;
-  imgSrc: string;
-  altTxt: string | undefined;
-}
-
-export interface Product {
-  node: {
-    id: string;
-    title: string;
-    images: {
-      edges: [
-        {
-          node: {
-            altText: string | undefined;
-            src: string;
-          };
-        }
-      ];
-    };
-  };
-}
+import type { CollectionObj, Collection, ProductObj, Product } from "~/types";
 
 export default function Home() {
   const [loading, setLoading] = useState(true);
